# Conflicts:
#	portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/AccountController.java

+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/DashboardController.java
@@ -4,46 +4,53 @@ import nz.ac.canterbury.seng302.portfolio.model.Project;
 import nz.ac.canterbury.seng302.portfolio.model.Sprint;
 import nz.ac.canterbury.seng302.portfolio.service.DashboardService;
 import nz.ac.canterbury.seng302.portfolio.service.SprintService;
+import nz.ac.canterbury.seng302.portfolio.service.UserAccountClientService;
+import nz.ac.canterbury.seng302.shared.identityprovider.AuthState;
+import nz.ac.canterbury.seng302.shared.identityprovider.ClaimDTO;
+import nz.ac.canterbury.seng302.shared.identityprovider.UserRole;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
-import java.sql.Date;
 import java.time.LocalDate;
-
+import java.util.Arrays;
 import java.util.List;
-
 
 @Controller
 public class DashboardController {
     @Autowired private DashboardService dashboardService;
+    @Autowired private UserAccountClientService userAccountClientService;
+    @Autowired private SprintService sprintService;
 
 
     /**
-     * Adds project list to model and opens dashboard.html
+     * Adds the project list to the Dashboard.
+     * Gets the list of the current user's roles
+     * Opens dashboard.html
      * @param model
-     * @return
+     * @param principal - current user.
+     * @return name of the dashboard html file.
      */
     @GetMapping("/dashboard")
-    public String showProjectList(Model model) {
-        List<Project> listProjects = dashboardService.listAll();
-        // Sets a default project if there are no projects
-        if (listProjects.isEmpty()) {
-
-            LocalDate now = LocalDate.now();
-            Project defaultProject = new Project();
-            defaultProject.setProjectName("Project " + now.getYear()); // Project {year}
-            defaultProject.setStartDate(Date.valueOf(now)); // Current date
-
-            defaultProject.setEndDate(Date.valueOf(now.plusMonths(8))); // 8 months from start date
-            dashboardService.saveProject(defaultProject);
-            listProjects.add(defaultProject);
+    public String showProjectList( @AuthenticationPrincipal AuthState principal,
+                                   Model model) {
+        List<Project> listProjects = null;
+        try {
+            listProjects = dashboardService.getAllProjects();
+            model.addAttribute("listProjects", listProjects);
+            // Add the list of the current users roles to model so they can be used in dashboard.html with thymeleaf.
+            model.addAttribute("roles", userAccountClientService.getUserRole(principal));
+            model.addAttribute("user", userAccountClientService.getUser(principal));
+            return "dashboard";
+        } catch (Exception e) {
+            return "error";
         }
-        model.addAttribute("listProjects", listProjects);
-        return "dashboard";
     }
 
     /**
@@ -52,9 +59,13 @@ public class DashboardController {
      * @return
      */
     @GetMapping("/dashboard/newProject")
-    public String showNewForm(Model model) {
+    public String showNewForm(Model model, @AuthenticationPrincipal AuthState principal) {
+        if (retrieveUserRoles(principal)) return "redirect:/dashboard";
+        dashboardService.setPreviousFeature("Create");
         model.addAttribute("project", new Project());
         model.addAttribute("pageTitle", "Add New Project");
+        model.addAttribute("submissionName", "Create");
+        model.addAttribute("user", userAccountClientService.getUser(principal));
         return "projectForm";
     }
 
@@ -64,9 +75,24 @@ public class DashboardController {
      * @return
      */
     @PostMapping("/dashboard/saveProject")
-    public String saveProject(Project project) {
-        dashboardService.saveProject(project);
-        return "redirect:/dashboard";
+    public String saveProject(
+            Project project,
+            Model model,
+            RedirectAttributes ra,
+            @AuthenticationPrincipal AuthState principal) {
+        if (retrieveUserRoles(principal)) return "redirect:/dashboard";
+        try {
+            String msgString;
+            dashboardService.saveProject(project);
+            msgString = dashboardService.createSuccessMessage(project);
+            if (msgString == null) {
+                throw new Exception("Error generating success message");
+            }
+            ra.addFlashAttribute("messageSuccess", msgString);
+            return "redirect:/dashboard";
+        } catch (Exception e) {
+            return "error";
+        }
     }
 
     /**
@@ -76,11 +102,25 @@ public class DashboardController {
      * @return
      */
     @GetMapping("/dashboard/editProject/{projectId}")
-    public String showEditForm(@PathVariable(value = "projectId") int projectId, Model model) {
-        Project project  = dashboardService.getProject(projectId);
-        model.addAttribute("project", project);
-        model.addAttribute("pageTitle", "Edit Project (Name: " + projectId + ")");
-        return "projectForm";
+    public String showEditForm(
+        @PathVariable(
+        value = "projectId") int projectId,
+        Model model,
+        RedirectAttributes ra,
+        @AuthenticationPrincipal AuthState principal) {
+        if (retrieveUserRoles(principal)) return "redirect:/dashboard";
+        try {
+            dashboardService.setPreviousFeature("Edit");
+            Project project  = dashboardService.getProject(projectId);
+            model.addAttribute("project", project);
+            model.addAttribute("pageTitle", "Edit Project: " + project.getProjectName());
+            model.addAttribute("submissionName", "Save");
+            model.addAttribute("user", userAccountClientService.getUser(principal));
+            return "projectForm";
+        } catch (NullPointerException e) {
+            ra.addFlashAttribute("messageDanger", "No Project Found");
+            return "redirect:/dashboard";
+        }
     }
 
     /**
@@ -90,21 +130,25 @@ public class DashboardController {
      * @throws Exception If project is not found in the database
      */
     @GetMapping("/dashboard/deleteProject/{projectId}")
-    public String showEditForm(@PathVariable("projectId") int projectId) throws Exception {
-        SprintService sprintService = new SprintService();
+    public String deleteProject(
+        @PathVariable("projectId") int projectId,
+        Model model,
+        @AuthenticationPrincipal AuthState principal) {
+        if (retrieveUserRoles(principal)) return "redirect:/dashboard";
         try {
-            List<Sprint> sprintList = sprintService.getAllSprints();
-            for (Sprint i: sprintList) {
-                if (i.getProject().getProjectId() == projectId) {
-                    sprintService.deleteSprint(i.getSprintId());
-                }
-            }
+            sprintService.deleteAllSprints(projectId);
+            dashboardService.deleteProject(projectId);
+            return "redirect:/dashboard";
         } catch (Exception e) {
-            e.printStackTrace();
+            return "error";
         }
-
-        dashboardService.deleteProject(projectId);
-        return "redirect:/dashboard";
     }
 
+    static boolean retrieveUserRoles(@AuthenticationPrincipal AuthState principal) {
+        List<String> userRoles = Arrays.asList(principal.getClaimsList().stream().filter(claim -> claim.getType().equals("role")).findFirst().map(ClaimDTO::getValue).orElse("NOT FOUND").split(","));
+        if (!(userRoles.contains(UserRole.TEACHER.name()) || userRoles.contains(UserRole.COURSE_ADMINISTRATOR.name()))) {
+            return true;
+        }
+        return false;
+    }
 }
