+++ portfolio/src/main/java/nz/ac/canterbury/seng302/portfolio/controller/SprintController.java
@@ -4,29 +4,57 @@ import nz.ac.canterbury.seng302.portfolio.model.Project;
 import nz.ac.canterbury.seng302.portfolio.model.Sprint;
 import nz.ac.canterbury.seng302.portfolio.service.ProjectService;
 import nz.ac.canterbury.seng302.portfolio.service.SprintService;
+import nz.ac.canterbury.seng302.portfolio.service.UserAccountClientService;
+import nz.ac.canterbury.seng302.shared.identityprovider.AuthState;
+import nz.ac.canterbury.seng302.shared.identityprovider.ClaimDTO;
+
+import nz.ac.canterbury.seng302.shared.identityprovider.UserRole;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.servlet.mvc.support.RedirectAttributes;
 
-import java.sql.Date;
-import java.time.LocalDate;
+import java.util.Arrays;
 import java.util.List;
 
 @Controller
 public class SprintController {
     @Autowired private SprintService sprintService;
     @Autowired private ProjectService projectService;
+    @Autowired private UserAccountClientService userAccountClientService;
 
+    /**
+     * Add project details, sprints, and current user roles (to determine access to add, edit, delete sprints)
+     * to the individual project pages.
+     * @param projectId - ID of the project selected to view.
+     * @param principal - Current User.
+     * @param model
+     * @return - name of the html page to display
+     * @throws Exception
+     */
     @GetMapping("/project/{projectId}")
-    public String showSprintList(@PathVariable("projectId") int projectId, Model model) throws Exception {
-        List<Sprint> listSprints = sprintService.getSprintByProject(projectId);
-        Project project = projectService.getProjectById(projectId);
-        model.addAttribute("listSprints", listSprints);
-        model.addAttribute("project", project);
-        return "project";
+    public String showSprintList(
+            @PathVariable("projectId") int projectId,
+            @AuthenticationPrincipal AuthState principal,
+            Model model,
+            RedirectAttributes ra) {
+        try {
+            List<Sprint> listSprints = sprintService.getSprintByProject(projectId);
+            Project project = projectService.getProjectById(projectId);
+            model.addAttribute("listSprints", listSprints);
+            model.addAttribute("project", project);
+            model.addAttribute("roles",
+                    userAccountClientService.getUserRole(principal));
+            model.addAttribute("user", userAccountClientService.getUser(principal));
+            return "project";
+        } catch (Exception e) {
+            ra.addFlashAttribute("messageDanger", e.getMessage());
+            return "redirect:/dashboard";
+        }
     }
 
     /**
@@ -35,26 +63,25 @@ public class SprintController {
      * @return
      */
     @GetMapping("/project/{projectId}/newSprint")
-    public String newSprint(Model model, @PathVariable ("projectId") int projectId) throws Exception {
-        int sprintNo = sprintService.countByProjectId(projectId) + 1;
-        Sprint newSprint = new Sprint();
-        List<Sprint> listSprints = sprintService.getSprintByProject(projectId);
-        if(listSprints.size() == 0) {
-
+    public String newSprint(
+            Model model,
+            @PathVariable ("projectId") int projectId,
+            @AuthenticationPrincipal AuthState principal,
+            RedirectAttributes ra){
+        if (DashboardController.retrieveUserRoles(principal)) return "redirect:/dashboard";
+        try {
             Project currentProject = projectService.getProjectById(projectId);
-            LocalDate startDate = currentProject.getStartDate().toLocalDate();
-            newSprint.setStartDate(Date.valueOf(startDate));
-            newSprint.setEndDate(Date.valueOf(startDate.plusWeeks(3)));
+            Sprint newSprint = sprintService.getNewSprint(currentProject);
+            model.addAttribute("pageTitle", "Add New Sprint");
+            model.addAttribute("sprint", newSprint);
+            model.addAttribute("project", currentProject);
+            model.addAttribute("user", userAccountClientService.getUser(principal));
+            return "sprintForm";
 
-        } else {
-            Sprint last_sprint = listSprints.get(listSprints.size() - 1);
-            LocalDate startDate = last_sprint.getEndDate().toLocalDate();
-            newSprint.setStartDate(Date.valueOf(startDate));
-            newSprint.setEndDate(Date.valueOf(startDate.plusWeeks(3)));
+        } catch (Exception e) {
+            ra.addFlashAttribute("messageDanger", e.getMessage());
+            return "redirect:/project/{projectId}";
         }
-        newSprint.setSprintName("Sprint " + sprintNo);
-        model.addAttribute("sprint", newSprint);
-        return "sprintForm";
     }
 
     /**
@@ -63,29 +90,53 @@ public class SprintController {
      * @return
      */
     @PostMapping("/project/{projectId}/saveSprint")
-    public String saveSprint(@PathVariable int projectId, Sprint sprint) throws Exception {
-        sprint.setProject(projectService.getProjectById(projectId));
-        sprintService.saveSprint(sprint);
-        return "redirect:/project/{projectId}";
-    }
+    public String saveSprint(
+        @PathVariable int projectId,
+        Sprint sprint,
+        @AuthenticationPrincipal AuthState principal,
+        RedirectAttributes ra) {
+        if (DashboardController.retrieveUserRoles(principal)) return "redirect:/dashboard";
+        try {
+                sprint.setProject(projectService.getProjectById(projectId));
+                String message = sprintService.saveSprint(sprint);
+                ra.addFlashAttribute("messageSuccess", message);
+            } catch (Exception e) {
+                ra.addFlashAttribute("messageDanger", e.getMessage());
+            }
+            return "redirect:/project/{projectId}";
+        }
 
 
-    /**
-     * Directs to page for editing a sprint
-     * @param sprintId ID for sprint being edited
-     * @param model
-     * @return
-     */
+
+        /**
+         * Directs to page for editing a sprint
+         * @param sprintId ID for sprint being edited
+         * @param model
+         * @return
+         */
     /*make sure to update project.html for path*/
     @GetMapping("/project/{projectId}/editSprint/{sprintId}")
-    public String sprintEditForm(@PathVariable("sprintId") int sprintId, @PathVariable("projectId") int projectId, Model model) throws Exception {
-        Sprint sprint = sprintService.getSprint(sprintId);
-        model.addAttribute("sprint", sprint);
-        model.addAttribute("pageTitle", "Edit Sprint (Name: " + sprintId + ")");
-        return "sprintForm";
+    public String sprintEditForm(
+            @PathVariable("sprintId") int sprintId,
+            @PathVariable("projectId") int projectId,
+            Model model,
+            @AuthenticationPrincipal AuthState principal,
+            RedirectAttributes ra){
+        if (DashboardController.retrieveUserRoles(principal)) return "redirect:/dashboard";
+        try {
+                Project currentProject = projectService.getProjectById(projectId);
+                Sprint sprint = sprintService.getSprint(sprintId);
+                model.addAttribute("sprint", sprint);
+                model.addAttribute("project", currentProject);
+                model.addAttribute("pageTitle", "Edit Sprint: " + sprint.getSprintName());
+            model.addAttribute("user", userAccountClientService.getUser(principal));
+                return "sprintForm";
+            } catch (Exception e) {
+                ra.addFlashAttribute("messageDanger", e.getMessage());
+                return "redirect:/project/{projectId}";
+            }
+        }
 
-
-    }
 
     /**
      * Deletes a sprint and redirects back to project page
@@ -93,12 +144,21 @@ public class SprintController {
      * @param model
      * @return
      */
-    /* Deleting sprints */
     @GetMapping("/project/{projectId}/deleteSprint/{sprintId}")
-    public String deleteSprint(@PathVariable("sprintId") int sprintId, Model model, @PathVariable int projectId){
-        sprintService.deleteSprint(sprintId);
+    public String deleteSprint(
+        @PathVariable("sprintId") int sprintId,
+        Model model,
+        @PathVariable int projectId,
+        @AuthenticationPrincipal AuthState principal,
+        RedirectAttributes ra){
+        if (DashboardController.retrieveUserRoles(principal)) return "redirect:/dashboard";
+        try {
+            String message = sprintService.deleteSprint(sprintId);
+            ra.addFlashAttribute("messageSuccess", message);
+        } catch (Exception e) {
+            ra.addFlashAttribute("messageDanger", e.getMessage());
+        }
         sprintService.updateSprintNames(sprintService.getSprintByProject(projectId));
-        sprintService.getSprintByProject(projectId);
         List<Sprint> listSprints = sprintService.getSprintByProject(projectId);
         model.addAttribute("listSprints", listSprints);
         return "redirect:/project/{projectId}";
